backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pytest
pymongo
google-generativeai
